Keep timings class intact so diff is actually readable
