+++ Renderer/ParticleRenderer/Renderers/RenderCables.cs
@@ -441,7 +441,7 @@ namespace ValveResourceFormat.Renderer.Particles.Renderers
             GL.Disable(EnableCap.Blend);
             GL.DepthMask(true);
 
-            PerfStats.Active?.CountParticleDraw();
+            Counters.Active.CountParticleDraw();
             GL.DrawElements(PrimitiveType.Triangles, indexCount, DrawElementsType.UnsignedInt, 0);
 
             material.PostRender();
