+++ Renderer/ParticleRenderer/Renderers/RenderSprites.cs
@@ -421,7 +421,7 @@ namespace ValveResourceFormat.Renderer.Particles.Renderers
             shader.SetUniform1("uTextureChannels", (int)textureChannels);
 
             // DRAW
-            PerfStats.Active?.CountParticleDraw();
+            Counters.Active.CountParticleDraw();
             GL.DrawElements(PrimitiveType.Triangles, particleBag.Count * 6, DrawElementsType.UnsignedShort, 0);
 
             GL.UseProgram(0);
